Move common methods to library
